added test currentIndexCity
